+++ Bancos_Contabilidad/Bancos_Contabilidad/frmTipoCuenta.cs
@@ -17,6 +17,9 @@ namespace Bancos_Contabilidad
         public frmTipoCuenta()
         {
             InitializeComponent();
+            btnGuardar.Enabled = txtNombre.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = true;
+            btnNuevo.Enabled = btnEditar.Enabled = btnEliminar.Enabled = false;
+            estado = "insertar";
         }
         public void limpiar()
         {
@@ -45,13 +48,15 @@ namespace Bancos_Contabilidad
         private void btnEditar_Click(object sender, EventArgs e)
         {
             estado = "editar";
-            btnGuardar.Enabled = btnCancelar.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = true;
+            btnGuardar.Enabled = btnCancelar.Enabled = txtNombre.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = true;
+            btnNuevo.Enabled = btnEliminar.Enabled = false;
         }
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
             estado = "eliminar";
             btnGuardar.Enabled = btnCancelar.Enabled = true;
+            btnNuevo.Enabled = btnEditar.Enabled = false;
         }
 
         private void btnGuardar_Click(object sender, EventArgs e)
@@ -65,6 +70,10 @@ namespace Bancos_Contabilidad
                 p.Status = "1";
                 CapaNegocio cn = new CapaNegocio();
                 cn.insertCuentaContable(p);
+                limpiar();
+                btnNuevo.Enabled = true;
+                btnEditar.Enabled = btnEliminar.Enabled = btnGuardar.Enabled = btnCancelar.Enabled = txtNombre.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = false;
+
             }
             else if (estado == "editar")
             {
@@ -76,19 +85,29 @@ namespace Bancos_Contabilidad
                 CapaNegocio cn = new CapaNegocio();
                 p.Status = "1";
                 cn.actualizarCuentaContable(p);
+                btnEditar.Enabled = btnEliminar.Enabled = btnNuevo.Enabled = true;
+                btnGuardar.Enabled = btnCancelar.Enabled = txtNombre.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = false;
             }
             else if (estado == "eliminar")
              {
-                 CapaEntidad.cuentacontable p = new CapaEntidad.cuentacontable();
-                 p.ID1 = id;
-                 CapaNegocio cn = new CapaNegocio();
-                 cn.eliminarCuentaContable(p);
+                if (MessageBox.Show("Desea eliminar el registro", "Eliminacion", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    CapaEntidad.cuentacontable p = new CapaEntidad.cuentacontable();
+                    p.ID1 = id;
+                    CapaNegocio cn = new CapaNegocio();
+                    cn.eliminarCuentaContable(p);
+                    limpiar();
+                    btnNuevo.Enabled = true;
+                    btnEditar.Enabled = btnEliminar.Enabled = btnGuardar.Enabled = btnCancelar.Enabled = txtNombre.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = false;
+                }
+                
              }
         }
 
         private void btnCancelar_Click(object sender, EventArgs e)
         {
-            
+            btnNuevo.Enabled = btnEditar.Enabled = btnEliminar.Enabled = true;
+            btnEditar.Enabled = btnGuardar.Enabled = btnCancelar.Enabled = txtNombre.Enabled = cmbEstadoFinanciero.Enabled = cmbClasificacion.Enabled = false;
         }
     }
 }
